Create tribute page for Dr. A.P.J. Abdul Kalam
Added HTML structure and content for a tribute page.
User Stories:

Your tribute page should have a main element with a corresponding id of main, which contains all other elements
You should see an element with an id of title, which contains a string (i.e. text), that describes the subject of the tribute page (e.g. "Dr. Norman Borlaug")
You should see either a figure or a div element with an id of img-div
Within the #img-div element, you should see an img element with a corresponding id="image"
Within the #img-div element, you should see an element with a corresponding id="img-caption" that contains textual content describing the image shown in #img-div
You should see an element with a corresponding id="tribute-info", which contains textual content describing the subject of the tribute page
You should see an a element with a corresponding id="tribute-link", which links to an outside site, that contains additional information about the subject of the tribute page. HINT: You must give your element an attribute of target and set it to _blank in order for your link to open in a new tab
Your #image should use max-width and height properties to resize responsively, relative to the width of its parent element, without exceeding its original size
Your img element should be centered within its parent element